Refactor Emitter to comply with the new verbosity levels behaviour.
Note that Emitter final behaviour (as described in the Spec) will only be achieved after more several PRs, I added XXX notes for changes that will happen in the short term future.

This branch is for sure the most "conceptual" one in the "verbosity levels" series. The changes in the code are not many, but the tests are seriously affected (and will continue to mutate when other PRs land). My recommendation is to analyze the changes in the code using [the Spec itself](https://docs.google.com/document/d/1pJDT-NE3jhkfVvDw6ba0Z3adnaQnGRi9slC9jIOqrZ4/edit) for better reasoning..