Add test
This test generates and destroys 514 objects from the
_PipeReaderThreader class. Since the limit value in FDs for
select is 1024 in both Linux and BSD/MacOS, and each object
creates two FDs, if there is any kind of problem when freeing
the FDs this test will fail. Only if the FDs are correctly
freed after each destroy this test will succeed.