(messages) Show messages from subprocess in an ephemeral way when in brief mode.
Also added an example to show this (and improved the whole examples machinery to accept parameters, to make it easier to use).

Added a couple of small changes because pyright/pylint, and will split the big `messages.py` in a following branch (without semantic changes in it).

Documentation updated.

No change needed from the applications using the library.