docs: consolidate craft-cli and starbase docs
This commit brings craft-cli's documentation into the format and build
process that Starbase imposes. The specific Makefile and requirements
files in docs/ are removed in favor of the tox 'build-docs' and
'autobuild-docs' that use the dependencies from pyproject.toml.

sphinx-build is called *without* treating warnings as errors, because
the auto-generated doc process generates a bunch of warnings about
circular imports and duplicate declarations that should be fixed at
a later date.

Also remove .readthedocs.yml in favor of the starbase-provided
.readthedocs.yaml.