messages: replace tabs with spaces in open_stream() (#151)
The tabs complicate the internal book-keeping because "\t12" has 3
characters but is printed to the terminal with a varying number of
spaces (depending on the user's tabwidth). For the pipe reader used
in open_stream(), this issue manifested as a bug where lines containing
tabs would result in multiple lines being printed (instead of just the
last one), which breaks the "ephemeral" aspect of the stream.

Note that this discovery implies that we should probably revisit this
tab issue in more places; in particular, the Printer does multiple
comparisons between the len() of the text and the terminal width, so
if the text has tabs that comparison will be incorrect. However, this
commit aims only to fix the (unreleased) behavior of showing subprocess
output as ephemeral messages inside open_stream().