tools: call 'coverage' directly
The problem: when using pytest-cov to integrate coverage with the test
run, there is a 'race' between pytest-cov and the pytest plugin provided
by craft-cli (the emitter). When craft-cli's plugin is loaded first, it
necessitates importing the main 'craft_cli' package, which in turn
imports dispatcher.py, messages.py, etc.

As a result, all the top-level lines (class and function definitions,
global-variables, etc) are loaded _before_ coverage starts running,
meaning that they are marked as 'hits' because they are never "loaded
again" during the tests. Instead, use coverage specifically so that
it's always set-up before the whole importing starts.