docs: document class attributes separately
Instead of keeping their docs in the main class docstring, use per-
attribute docstrings to improve the readibility and "searchability" of
the items. Some minor tweaks and adjustments to the text itself, but
nothing too major.

Note that we also drop the "sphinx.ext.more_autodoc" extension because
it isn't adding anything (that I can tell) other than an annoying double
list of init parameters that I couldn't get rid of.