printer: always flush when logging messages
With this commit, every write() call to the Printer's logfile is
followed by an explicit call to flush(). This is in line with the
standard library's FileHandler behavior, and lets us correctly write
to file messages logged from child processes (multiprocessing). Without
this, the files are written() to but the child process dies without
flushing them to disk, so they are lost.