emitter: add "streaming brief" feature
This commit adds a new feature to the Emitter, tentatively called
"streaming brief". It is initially disabled, and can be enabled in
the Emitter's init() call.

Once enabled, the feature will leverage the "multi-action" nature of
progress() calls to "stream", in a single line, info-level log messages
and text written to the open_stream()'s pipe when in BRIEF mode. For
example, the following sequence of calls (in BRIEF mode):

```
emit.progress("Starting stage 1", permanent=False)
...
log.info("Doing first step")
...
log.info("Doing second step")
...
emit.progress("Finished stage 1", permanent=True)
```

... will cause the two 'info' messages to "stream" on the terminal,
prefixed by "Starting stage 1 ::" to indicate that they are related to
that progress message. All these three messages are ephemeral which
means that in the end only the final progress() call will be visible
on the terminal.

Fixes #165