tools: include integration tests in coverage
It makes sense to include the integration tests in the total coverage
metric because in craft-cli the integration tests are more about
testing the actual public api end-to-end. Often, these are the tests
that better reflect the real-world use of the package and the tests
that have a better chance of surviving true internal refactorings
without changes.