fix(dispatcher): revert a change in return type annotation
Commit b0698f9 changed the type annotation for the return value of
BaseCommand.run() from `Optional[int]` to the semantically-equivalent
preferred notation `int | None`.

However, this is a breaking change in the case of subclasses that
override this method, return only `None` and use the `override`
decorator from non-latest versions of the `override` package, as those
versions do not realize that `None` is a valid subset of `int | None`.

Although the fix in client code is simple, the original commit was not
meant to break client code at all. Therefore, revert this specific
change until we are ready to release a new major version.