fix: handle errors when decoding bytes from subprocesses
Since these bytes come from the output of unknown processes, there is no
guarantee that they are valid utf-8 text. We also have no way to know
what encoding they have (if any), so the best we can do is replace the
invalid bytes with the usual unicode marker for that.

Fixes #221