feat: add CraftCommandError
This new error class has a ``stderr`` attribute that holds the error output
of an executed command. As the class' docstring indicates, this new attribute
is halfway between the ``message`` and the ``details`` parameters, verbosity
-wise. It's meant to be used in cases where this error output is expected to
be useful enough to the end user to make up for the extra text, even in
"brief"-mode executions.

Fixes #260